third_party/blink/web_tests/wpt_internal/css/css-anchor-position/parsing/position-try-order-parsing.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Anchor Positioning Test: Parsing of position-try-order</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/9196#issuecomment-1912788896">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<div id="target"></div>
<script>
  test_valid_value("position-try-order", "initial");
  test_valid_value("position-try-order", "inherit");
  test_valid_value("position-try-order", "unset");
  test_valid_value("position-try-order", "revert");
  test_valid_value("position-try-order", "normal");
  test_valid_value("position-try-order", "most-width");
  test_valid_value("position-try-order", "most-height");
  test_valid_value("position-try-order", "most-block-size");
  test_valid_value("position-try-order", "most-inline-size");

  test_invalid_value("position-try-order", "normal most-inline-size");
  test_invalid_value("position-try-order", "most-block-size most-inline-size");
  test_invalid_value("position-try-order", "most-block-size, most-inline-size");
</script>
